% Auto-generated: do not edit by hand
\name{''Testing}

\alias{''Testing}

\title{Testing component}

\description{
Component description
}

\usage{
''Testing(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. children of the component.}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. className for styling and other uses.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

