\name{abilityTest}
\alias{abilityTest}
\docType{data}
\title{Ability Test}
\description{
Simulated data for an hypothetical test aimed to measure a cognitive ability for children.
}
\usage{data("abilityTest")}
\details{
The column \code{score} reports the raw score (range: 0-30), while \code{errors} are the counts of the errors committed.
}
\examples{
data("abilityTest")
str(abilityTest)
}
\keyword{datasets}
