% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{add 1 to a number}
\usage{
add_one(number)
}
\arguments{
\item{number}{any number}
}
\value{
number added by one.
}
\description{
This function takes any numerical input and adds 1 to it.
}
\examples{
add_one(5)
}
