% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{estimate_pi}
\alias{estimate_pi}
\title{Estimate pi by random draws}
\usage{
estimate_pi(n)
}
\arguments{
\item{n}{The number of draws to make}
}
\description{
Estimate pi by making random draws and doing rejection sampling.
}
\examples{
estimate_pi(10)
# more accurate with bigger n.
estimate_pi(100000)
pi
}

