% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplits.R
\name{new_func}
\alias{new_func}
\title{Title}
\usage{
new_func(x, y)
}
\arguments{
\item{x}{integer 1}

\item{y}{integer 2}
}
\value{
a sum of a number and cube of a number
}
\description{
Title
}
\examples{
new_func(1,2)
}
