\name{normTab}
\alias{normTab}
\docType{data}
\title{Normative Table}
\description{
Example of normative table for an hypothetical test aimed to evaluate children of primary school (6-10 years).
}
\usage{data("normTab")}
\details{
Colnames represents age groups. Rownames represents achievable scores expanded on the Wechsler scale (mean: 10, standard deviation: 3).
Tabulated data are raw scores corresponding to each standard score for each age group.
}
\examples{
data("normTab")
str(normTab)
}
\keyword{datasets}
