% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_interviews.R
\docType{data}
\name{rinker_interviews}
\alias{rinker_interviews}
\title{Interviews}
\format{A data frame with 427 rows and 13 variables}
\usage{
data(rinker_interviews)
}
\description{
A dataset containing interview data.
}
\details{
\itemize{
  \item Observation. Obs #.
  \item Person. Person.
  \item ID. Person ID.
  \item Role. Person Role.
  \item Gender. Person Geners.
  \item Race. Person Race.
  \item Question_Number. Question ID.
  \item Question. Question text.
  \item Word_Count. Word count.
  \item element_id. Row number (tot).
  \item sentence_id. Sentence id within element_id.
  \item Dialogue. The discourse.
  \item Complete_Row_ID. Rows for where there was text.
}
}
\keyword{datasets}

