% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_interviews_formality.R
\docType{data}
\name{rinker_interviews_formality}
\alias{rinker_interviews_formality}
\title{Interview Formality}
\format{A data frame with 1 rows and 13 variables}
\usage{
data(rinker_interviews_formality)
}
\description{
A dataset containing a formality object for interview.
}
\details{
\itemize{
  \item all. The grouping variable.
  \item noun. Noun counts.
  \item preposition. Preposition counts.
  \item adjective. Adjective counts.
  \item article. Article counts.
  \item verb. Verb counts.
  \item pronoun. Pronoun counts.
  \item adverb. Adverb counts.
  \item interjection. Interjection counts.
  \item formal. Formal counts.
  \item contextual. Contextual counts.
  \item n. Total words aalyzed.
  \item F. Formality score.
}
}
\keyword{datasets}

