% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_interviews_hclust.R
\docType{data}
\name{rinker_interviews_hclust}
\alias{rinker_interviews_hclust}
\title{Interview Hierarchical Cluster}
\format{A list with 7 elements}
\usage{
data(rinker_interviews_hclust)
}
\description{
A dataset containing a \code{hierarchical_cluster} list.
}
\keyword{datasets}

