% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_interviews_sentiment.R
\docType{data}
\name{rinker_interviews_sentiment}
\alias{rinker_interviews_sentiment}
\title{Interview Sentiment}
\format{A data frame with 427 rows and 4 variables}
\usage{
data(rinker_interviews_sentiment)
}
\description{
A dataset containing sentiment data.
}
\keyword{datasets}

