% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_mech_turk_emotion.R
\docType{data}
\name{rinker_mech_turk_emotion}
\alias{rinker_mech_turk_emotion}
\title{Mechanic Turk Emotions for Interview}
\format{A data frame with 327 rows and 5 variables}
\usage{
data(rinker_mech_turk_emotion)
}
\description{
A dataset containing Mechanical Turk emotions rations for interviews.
}
\details{
\itemize{
  \item Coder_1. Coder's responses.
  \item Coder_2. Coder's responses.
  \item Coder_3. Coder's responses.
  \item Coder_4. Coder's responses.
  \item Coder_5. Coder's responses.
}
}
\keyword{datasets}

