% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_observations_hclust.R
\docType{data}
\name{rinker_observations_hclust}
\alias{rinker_observations_hclust}
\title{Observation Hierarchical Cluster}
\format{A list with 7 elements}
\usage{
data(rinker_observations_hclust)
}
\description{
A dataset containing a \code{hierarchical_cluster} list.
}
\keyword{datasets}

