% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinker_observations_sentiment.R
\docType{data}
\name{rinker_observations_sentiment}
\alias{rinker_observations_sentiment}
\title{Observation Sentiment}
\format{A data frame with 6141 rows and 4 variables}
\usage{
data(rinker_observations_sentiment)
}
\description{
A dataset containing sentiment data.
}
\keyword{datasets}

