% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_line_dat.R
\docType{data}
\name{sentiment_line_dat}
\alias{sentiment_line_dat}
\title{Data for sentiment trend line}
\format{A data frame with 327 rows and 36 variables}
\usage{
data(sentiment_line_dat)
}
\description{
A dataset containing sentiment related data
}
\details{
\itemize{ 
  \item Complete_Row_ID. id
  \item Sentiment_turk. Turk sentiment score
  \item Anger_turk. Turk anger score
  \item Anticipation_turk. Turk anticipation score
  \item Disgust_turk. Turk disgust score
  \item Fear_turk. Turk fear score
  \item Joy_turk. Turk joy score
  \item None_turk. Turk nada score
  \item Sadness_turk. Turk sadness score
  \item Surprise_turk. Turk surprise score
  \item Trust_turk. Turk trust score
  \item Observation. Observation number id
  \item Person. Teacher being interviewed
  \item ID. An id
  \item Role. Teacher or student
  \item Gender. Gender of teacher
  \item Race. Race of teacher
  \item Question_Number. Which question id
  \item Question. The question
  \item Word_Count. Words in response
  \item element_id. Which turn of talk
  \item sentence_id. Sentence id within turn of talk
  \item Dialogue. The speech
  \item Sentiment_comp. Computer (Stanford) sentiment score
  \item Anger_comp. Computer anger score
  \item Anticipation_comp. Computer anticipation score
  \item Disgust_comp. Computer disgust score
  \item Fear_comp. Computer fear score
  \item Joy_comp. Computer joy score
  \item Sadness_comp. Computer sadness score
  \item Surprise_comp. Computer surprise score
  \item Trust_comp. Computer trust score
  \item Negative_comp. Computer negative score
  \item Positive_comp. Computer positive score
  \item narative_duration. Sentence id number within person
  \item Question2. Better formatted question for plotting
}
}
\keyword{datasets}

