% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiread.R, R/testing-package.R
\docType{data}
\name{sentiread}
\alias{sentiread}
\title{Sentiment and Readability}
\format{A data frame with 327 rows and 42 variables}
\usage{
data(sentiread)

data(sentiread)
}
\description{
A dataset containing various sentiment and redability variables for interviews.

A dataset containing
}
\details{
\itemize{ 
  \item Complete_Row_ID. id
  \item Sentiment_turk. Turk sentiment score
  \item Anger_turk. Turk anger score
  \item Anticipation_turk. Turk anticipation score
  \item Disgust_turk. Turk disgust score
  \item Fear_turk. Turk fear score
  \item Joy_turk. Turk joy score
  \item None_turk. Turk nada score
  \item Sadness_turk. Turk sadness score
  \item Surprise_turk. Turk surprise score
  \item Trust_turk. Turk trust score
  \item Observation. Observation number id
  \item Person. Teacher being interviewed
  \item ID. An id
  \item Role. Teacher or student
  \item Gender. Gender of teacher
  \item Race. Race of teacher
  \item Question_Number. Which question id
  \item Question. The question
  \item Word_Count. Words in response
  \item element_id. Which turn of talk
  \item sentence_id. Sentence id within turn of talk
  \item Dialogue. The speech
  \item Sentiment_comp. Computer (Stanford) sentiment score
  \item Anger_comp. Computer anger score
  \item Anticipation_comp. Computer anticipation score
  \item Disgust_comp. Computer disgust score
  \item Fear_comp. Computer fear score
  \item Joy_comp. Computer joy score
  \item Sadness_comp. Computer sadness score
  \item Surprise_comp. Computer surprise score
  \item Trust_comp. Computer trust score
  \item Negative_comp. Computer negative score
  \item Positive_comp. Computer positive score
  \item narative_duration. Sentence id number within person
  \item Flesch_Kincaid. Readability statistic (specific)
  \item Gunning_Fog_Index. Readability statistic (specific)
  \item Coleman_Liau. Readability statistic (specific)
  \item SMOG. Readability statistic (specific)
  \item Automated_Readability_Index. Readability statistic (specific)
  \item Average_Grade_Level. Readability statistic (general/summary)
  \item ave_sentiment. Sentimentr sentiment score 
}

\itemize{ 
  \item Complete_Row_ID. 
  \item Sentiment_turk. 
  \item Anger_turk. 
  \item Anticipation_turk. 
  \item Disgust_turk. 
  \item Fear_turk. 
  \item Joy_turk. 
  \item None_turk. 
  \item Sadness_turk. 
  \item Surprise_turk. 
  \item Trust_turk. 
  \item Observation. 
  \item Person. 
  \item ID. 
  \item Role. 
  \item Gender. 
  \item Race. 
  \item Question_Number. 
  \item Question. 
  \item Word_Count. 
  \item element_id. 
  \item sentence_id. 
  \item Dialogue. 
  \item Sentiment_comp. 
  \item Anger_comp. 
  \item Anticipation_comp. 
  \item Disgust_comp. 
  \item Fear_comp. 
  \item Joy_comp. 
  \item Sadness_comp. 
  \item Surprise_comp. 
  \item Trust_comp. 
  \item Negative_comp. 
  \item Positive_comp. 
  \item narative_duration. 
  \item Flesch_Kincaid. 
  \item Gunning_Fog_Index. 
  \item Coleman_Liau. 
  \item SMOG. 
  \item Automated_Readability_Index. 
  \item Average_Grade_Level. 
  \item ave_sentiment. 
}
}
\keyword{datasets}

