% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplits.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x)
}
\arguments{
\item{x}{string data with comma}

\item{split}{where to split the string}
}
\value{
a character vector
}
\description{
Split a string
}
\examples{
x <- "alpha,beta,charlie,delta"
strsplit1(x, ".")

}
