% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ross.R
\name{test_ross}
\alias{test_ross}
\title{This means nothing}
\usage{
test_ross(x = 1, y = 2, z = 3)
}
\arguments{
\item{x}{this is a number}

\item{y}{this is another number}
}
\value{
Gives back 7.
}
\description{
This function is purely a test.
}
\examples{
test_ross(
x=2,
y=3
)
}
