\name{testDates}
\alias{testDates}
\docType{data}
\title{Test Dates}
\description{
Example of two-variable storage of age at the moment of administration of a psychometric test.
}
\usage{data("testDates")}
\details{
The column \code{born} is the date of birth. The column \code{test} is the date of test administration.
}
\examples{
data("testDates")
str(testDates)
}
\keyword{datasets}
