\name{parallel}
\alias{parallel}
\alias{print.parallel}
\alias{plot.parallel}
\title{Parallel Analysis}
\description{
To describe
}
\usage{
parallel(x, iter = 1000, ordinal = FALSE, method = "perm", alpha = 0.05,
         standard = FALSE, plot = TRUE, fn = eigen, ...)
}
\arguments{
    \item{x}{description}
    \item{iter}{description}
    \item{ordinal}{description}
    \item{method}{description}
    \item{alpha}{description}
    \item{standard}{description}
    \item{plot}{description}
    \item{fn}{description}
    \item{...}{description}
}
\details{
to do
}
\author{
Davide Massidda \email{davide.massidda@gmail.com}
}
\seealso{
\code{\link[psych]{fa.parallel}}
}
\examples{
# to do
}
